commit-72 Repository

+++ src/main/java/pl/yummy/domain/Address.java
@@ -9,7 +9,7 @@ import lombok.*;
 @ToString(of = {"addressId", "country", "city", "postalCode", "street"})
 public class Address {
 
-    Integer addressId;
+    Long addressId;
     String country;
     String city;
     String postalCode;
